Revert breaking windows changes

+++ freetype/src/pcf/pcfutil.h
@@ -37,15 +37,15 @@ FT_BEGIN_HEADER
 
   FT_LOCAL( void )
   BitOrderInvert( unsigned char*  buf,
-                  size_t          nbytes );
+                  int             nbytes );
 
   FT_LOCAL( void )
   TwoByteSwap( unsigned char*  buf,
-               size_t          nbytes );
+               int             nbytes );
 
   FT_LOCAL( void )
   FourByteSwap( unsigned char*  buf,
-                size_t          nbytes );
+                int             nbytes );
 
 FT_END_HEADER
